Add Tests for Description & Title components
